Extract validator from processor.go to a new validation.go.
This unfortunately exposes that there are no unit tests for this code directly (it's exercised
indirectly through processor tests).
